requirements.txt
web3==7.14.0
pandas==3.0.0
matplotlib==3.10.8
requests==2.32.5
python-dotenv==1.2.1
